fix: egit the return value of the iterator
