+++ crates/burn-communication/src/util.rs
@@ -1,8 +1,3 @@
-use tracing_core::{Level, LevelFilter};
-use tracing_subscriber::{
-    Layer, filter::filter_fn, layer::SubscriberExt, registry, util::SubscriberInitExt,
-};
-
 /// Utilities to help handle communication termination.
 pub async fn os_shutdown_signal() {
     let ctrl_c = async {
@@ -26,21 +21,4 @@ pub async fn os_shutdown_signal() {
         _ = ctrl_c => {},
         _ = terminate => {},
     }
-}
-
-pub(crate) fn init_logging() {
-    let layer = tracing_subscriber::fmt::layer()
-        .with_filter(LevelFilter::INFO)
-        .with_filter(filter_fn(|m| {
-            if let Some(path) = m.module_path() {
-                // The wgpu crate is logging too much, so we skip `info` level.
-                if path.starts_with("wgpu") && *m.level() >= Level::INFO {
-                    return false;
-                }
-            }
-            true
-        }));
-
-    // If we start multiple servers in the same process, this will fail, it's ok
-    let _ = registry().with(layer).try_init();
 }
