+++ crates/burn-communication/src/websocket/server.rs
@@ -1,9 +1,6 @@
 use std::net::SocketAddr;
 
-use crate::{
-    base::{CommunicationChannel, CommunicationError, Message, ProtocolServer},
-    util::init_logging,
-};
+use crate::base::{CommunicationChannel, CommunicationError, Message, ProtocolServer};
 use axum::{
     Router,
     extract::{
@@ -16,6 +13,27 @@ use futures::{
     SinkExt, StreamExt,
     stream::{SplitSink, SplitStream},
 };
+use tracing_core::{Level, LevelFilter};
+use tracing_subscriber::{
+    Layer, filter::filter_fn, layer::SubscriberExt, registry, util::SubscriberInitExt,
+};
+
+fn init_logging() {
+    let layer = tracing_subscriber::fmt::layer()
+        .with_filter(LevelFilter::INFO)
+        .with_filter(filter_fn(|m| {
+            if let Some(path) = m.module_path() {
+                // The wgpu crate is logging too much, so we skip `info` level.
+                if path.starts_with("wgpu") && *m.level() >= Level::INFO {
+                    return false;
+                }
+            }
+            true
+        }));
+
+    // If we start multiple servers in the same process, this will fail, it's ok
+    let _ = registry().with(layer).try_init();
+}
 
 #[derive(Clone, Debug)]
 pub struct WsServer {
